+++ pyvista/core/composite.py
@@ -12,7 +12,7 @@ from vtk import vtkMultiBlockDataSet
 
 import pyvista
 from pyvista.utilities import get_array, is_pyvista_dataset, wrap
-from .common import DataObject
+from .dataobject import DataObject
 from .filters import CompositeFilters
 
 log = logging.getLogger(__name__)
